Extract Rust command helpers and split the local store module
run_daylite_command builds the client and runs the token-locked core for
the three read-only project commands, which also removes their store
save: they never mutate the store, so the load-save round trip was a
no-op (holiday cache pruning still happens on every genuine save).
load_week_events is broken into employees_with_error,
fetch_week_for_employees, fetch_uncached_projects, and
assemble_week_events, replacing the first-pass/second-pass comments.
local_store.rs is split into local_store/types.rs and
local_store/persistence.rs with re-exports keeping all paths stable.
