+++ src-tauri/src/integrations/calendar/commands.rs
@@ -13,6 +13,7 @@ use super::events::{
     classify_event, map_absence_raw_events_for_week, resolve_event, sort_events_absences_first,
 };
 use super::types::{CalendarCellEvent, EmployeeWeekEvents, PendingEvent};
+use crate::integrations::local_store::{DayliteCache, LocalStore};
 
 /// Loads all calendar events for every configured employee for the given week.
 /// Returns one entry per employee that has a primary calendar configured.
@@ -30,26 +31,10 @@ pub async fn load_week_events(
     let week_start_date = NaiveDate::parse_from_str(&week_start, "%Y-%m-%d")
         .map_err(|_| format!("Ungültiges Wochenstartdatum: {week_start}"))?;
 
-    let (username, password) = match crate::integrations::zep::load_zep_credentials_from_keychain()
-    {
-        Ok(c) => (c.username, c.password),
+    let credentials = match crate::integrations::zep::load_zep_credentials_from_keychain() {
+        Ok(c) => c,
         Err(e) => {
-            // No credentials: return error for every employee with a calendar URL
-            let results: Vec<EmployeeWeekEvents> = store
-                .employee_settings
-                .iter()
-                .filter(|s| {
-                    s.zep_primary_calendar
-                        .as_deref()
-                        .map(|u| !u.is_empty())
-                        .unwrap_or(false)
-                })
-                .map(|s| EmployeeWeekEvents {
-                    employee_reference: s.daylite_contact_reference.clone(),
-                    events: vec![],
-                    error: Some(e.user_message.clone()),
-                })
-                .collect();
+            let results = employees_with_error(&store, &e.user_message);
             if results.is_empty() {
                 eprintln!(
                     "load_week_events: ZEP credentials unavailable and no primary calendars configured: {}",
@@ -65,8 +50,62 @@ pub async fn load_week_events(
         .build()
         .map_err(|e| format!("HTTP-Client konnte nicht erstellt werden: {e}"))?;
 
-    // First pass: fetch CalDAV events for all employees concurrently (primary + absence per
-    // employee are also concurrent via tokio::join!).
+    let (fetches, error_results) = fetch_week_for_employees(
+        &store,
+        &client,
+        &credentials.username,
+        &credentials.password,
+        week_start_date,
+    )
+    .await;
+
+    let api_results = fetch_uncached_projects(app, &store, &fetches).await;
+
+    Ok(assemble_week_events(
+        fetches,
+        error_results,
+        &store.daylite_cache,
+        &api_results,
+    ))
+}
+
+struct EmployeeFetch {
+    employee_reference: String,
+    pending: Vec<PendingEvent>,
+    absences: Vec<CalendarCellEvent>,
+}
+
+/// Error entries for every employee with a primary calendar, used when the whole
+/// load cannot proceed (e.g. missing ZEP credentials).
+fn employees_with_error(store: &LocalStore, message: &str) -> Vec<EmployeeWeekEvents> {
+    store
+        .employee_settings
+        .iter()
+        .filter(|s| {
+            s.zep_primary_calendar
+                .as_deref()
+                .map(|u| !u.is_empty())
+                .unwrap_or(false)
+        })
+        .map(|s| EmployeeWeekEvents {
+            employee_reference: s.daylite_contact_reference.clone(),
+            events: vec![],
+            error: Some(message.to_string()),
+        })
+        .collect()
+}
+
+/// Fetches the week's CalDAV events for all employees concurrently (primary and
+/// absence calendar per employee are also concurrent via `tokio::join!`) and
+/// classifies the results. Employees whose primary fetch failed are returned as
+/// ready-made error entries instead.
+async fn fetch_week_for_employees(
+    store: &LocalStore,
+    client: &reqwest::Client,
+    username: &str,
+    password: &str,
+    week_start: NaiveDate,
+) -> (Vec<EmployeeFetch>, Vec<EmployeeWeekEvents>) {
     let employee_futures: Vec<_> = store
         .employee_settings
         .iter()
@@ -85,26 +124,16 @@ pub async fn load_week_events(
 
             let employee_ref = setting.daylite_contact_reference.clone();
             let client = client.clone();
-            let username = username.clone();
-            let password = password.clone();
+            let username = username.to_string();
+            let password = password.to_string();
 
             Some(async move {
                 let (primary_result, absence_result) = tokio::join!(
-                    fetch_calendar_events(
-                        &client,
-                        &calendar_url,
-                        &username,
-                        &password,
-                        week_start_date
-                    ),
+                    fetch_calendar_events(&client, &calendar_url, &username, &password, week_start),
                     async {
                         match absence_url {
                             Some(ref url) => fetch_calendar_events(
-                                &client,
-                                url,
-                                &username,
-                                &password,
-                                week_start_date,
+                                &client, url, &username, &password, week_start,
                             )
                             .await
                             .ok(),
@@ -119,22 +148,23 @@ pub async fn load_week_events(
 
     let fetch_results = futures::future::join_all(employee_futures).await;
 
-    let mut pending_per_employee: Vec<(String, Vec<PendingEvent>, Vec<CalendarCellEvent>)> =
-        Vec::new();
-    let mut error_results: Vec<EmployeeWeekEvents> = Vec::new();
+    let mut fetches = Vec::new();
+    let mut error_results = Vec::new();
 
-    for (employee_ref, primary_result, absence_result) in fetch_results {
+    for (employee_reference, primary_result, absence_result) in fetch_results {
         match primary_result {
             Ok(raw_events) => {
-                let pending = raw_events.into_iter().map(classify_event).collect();
-                let absence_events = absence_result
-                    .map(|raw| map_absence_raw_events_for_week(raw, week_start_date))
-                    .unwrap_or_default();
-                pending_per_employee.push((employee_ref, pending, absence_events));
+                fetches.push(EmployeeFetch {
+                    employee_reference,
+                    pending: raw_events.into_iter().map(classify_event).collect(),
+                    absences: absence_result
+                        .map(|raw| map_absence_raw_events_for_week(raw, week_start))
+                        .unwrap_or_default(),
+                });
             }
             Err(error_msg) => {
                 error_results.push(EmployeeWeekEvents {
-                    employee_reference: employee_ref,
+                    employee_reference,
                     events: vec![],
                     error: Some(error_msg),
                 });
@@ -142,10 +172,19 @@ pub async fn load_week_events(
         }
     }
 
-    // Collect unique project refs that are not in the local Daylite cache.
+    (fetches, error_results)
+}
+
+/// Fetches every project referenced by an event but missing from the local Daylite
+/// cache (sequential, typically few).
+async fn fetch_uncached_projects(
+    app: tauri::AppHandle,
+    store: &LocalStore,
+    fetches: &[EmployeeFetch],
+) -> HashMap<String, Option<(String, String)>> {
     let mut missing_refs: HashSet<String> = HashSet::new();
-    for (_, pending_events, _) in &pending_per_employee {
-        for event in pending_events {
+    for fetch in fetches {
+        for event in &fetch.pending {
             if let Some(ref project_ref) = event.project_ref {
                 let in_cache = store
                     .daylite_cache
@@ -159,8 +198,7 @@ pub async fn load_week_events(
         }
     }
 
-    // Second pass: fetch missing projects from the Daylite API (sequential, typically few).
-    let mut api_results: HashMap<String, Option<(String, String)>> = HashMap::new();
+    let mut api_results = HashMap::new();
     for project_ref in missing_refs {
         let result = crate::integrations::daylite::projects::fetch_project_by_reference(
             app.clone(),
@@ -170,27 +208,35 @@ pub async fn load_week_events(
         api_results.insert(project_ref, result);
     }
 
-    // Build final results combining cache and API lookups.
+    api_results
+}
+
+fn assemble_week_events(
+    fetches: Vec<EmployeeFetch>,
+    error_results: Vec<EmployeeWeekEvents>,
+    cache: &DayliteCache,
+    api_results: &HashMap<String, Option<(String, String)>>,
+) -> Vec<EmployeeWeekEvents> {
     let mut results = error_results;
-    for (employee_ref, pending_events, absence_events) in pending_per_employee {
-        let mut events: Vec<CalendarCellEvent> = pending_events
+    for fetch in fetches {
+        let mut events: Vec<CalendarCellEvent> = fetch
+            .pending
             .into_iter()
-            .map(|p| resolve_event(p, &store.daylite_cache, &api_results))
+            .map(|p| resolve_event(p, cache, api_results))
             .collect();
-        events.extend(absence_events);
+        events.extend(fetch.absences);
         // Deduplicate by UID to guard against CalDAV servers redelivering the same event.
         let mut seen_uids = HashSet::new();
         events.retain(|e| seen_uids.insert(e.uid.clone()));
-        // Absence events are shown first within each day.
         sort_events_absences_first(&mut events);
         results.push(EmployeeWeekEvents {
-            employee_reference: employee_ref,
+            employee_reference: fetch.employee_reference,
             events,
             error: None,
         });
     }
 
-    Ok(results)
+    results
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, Type)]
