+++ src-tauri/src/integrations/local_store/mod.rs
@@ -1,0 +1,5 @@
+pub mod persistence;
+pub mod types;
+
+pub use persistence::*;
+pub use types::*;
